build template of singlebook page, move books array to own file
