+++ components/Books.js
@@ -1,0 +1,1879 @@
+const books = [
+    {
+      kind: "books#volumes",
+      totalItems: 1,
+      items: [
+        {
+          kind: "books#volume",
+          id: "JQGv8eqVMrMC",
+          etag: "XMN/JYILcs8",
+          selfLink: "https://www.googleapis.com/books/v1/volumes/JQGv8eqVMrMC",
+          volumeInfo: {
+            title: "To Kill a Mockingbird",
+            authors: ["Harper Lee"],
+            publisher: "Random House",
+            publishedDate: "2004",
+            description: "Theatre program.",
+            industryIdentifiers: [
+              {
+                type: "ISBN_13",
+                identifier: "9780099466734",
+              },
+              {
+                type: "ISBN_10",
+                identifier: "0099466732",
+              },
+            ],
+            readingModes: {
+              text: false,
+              image: false,
+            },
+            pageCount: 307,
+            printType: "BOOK",
+            categories: ["African Americans"],
+            maturityRating: "NOT_MATURE",
+            allowAnonLogging: false,
+            contentVersion: "0.13.1.0.preview.0",
+            panelizationSummary: {
+              containsEpubBubbles: false,
+              containsImageBubbles: false,
+            },
+            imageLinks: {
+              smallThumbnail:
+                "http://books.google.com/books/content?id=JQGv8eqVMrMC&printsec=frontcover&img=1&zoom=5&edge=curl&source=gbs_api",
+              thumbnail:
+                "http://books.google.com/books/content?id=JQGv8eqVMrMC&printsec=frontcover&img=1&zoom=1&edge=curl&source=gbs_api",
+            },
+            language: "en",
+            previewLink:
+              "http://books.google.co.uk/books?id=JQGv8eqVMrMC&printsec=frontcover&dq=isbn:9780099466734&hl=&cd=1&source=gbs_api",
+            infoLink:
+              "http://books.google.co.uk/books?id=JQGv8eqVMrMC&dq=isbn:9780099466734&hl=&source=gbs_api",
+            canonicalVolumeLink:
+              "https://books.google.com/books/about/To_Kill_a_Mockingbird.html?hl=&id=JQGv8eqVMrMC",
+          },
+          saleInfo: {
+            country: "GB",
+            saleability: "NOT_FOR_SALE",
+            isEbook: false,
+          },
+          accessInfo: {
+            country: "GB",
+            viewability: "PARTIAL",
+            embeddable: true,
+            publicDomain: false,
+            textToSpeechPermission: "ALLOWED",
+            epub: {
+              isAvailable: false,
+            },
+            pdf: {
+              isAvailable: false,
+            },
+            webReaderLink:
+              "http://play.google.com/books/reader?id=JQGv8eqVMrMC&hl=&printsec=frontcover&source=gbs_api",
+            accessViewStatus: "SAMPLE",
+            quoteSharingAllowed: false,
+          },
+          searchInfo: {
+            textSnippet:
+              "Theatre program. WINNER OF THE PULITZER PRIZE &#39;No one ever forgets this book&#39; Independent &#39;Shoot all the bluejays you want, if you can hit &#39;em, but remember it&#39;s a sin to kill a mockingbird.",
+          },
+        },
+      ],
+    },
+    {
+      kind: "books#volumes",
+      totalItems: 1,
+      items: [
+        {
+          kind: "books#volume",
+          id: "KJvuJnzbtJAC",
+          etag: "QUpVm4yhybo",
+          selfLink: "https://www.googleapis.com/books/v1/volumes/KJvuJnzbtJAC",
+          volumeInfo: {
+            title: "Great Expectations",
+            authors: ["Charles Dickens"],
+            publisher: "Penguin UK",
+            publishedDate: "2003-01-30",
+            description:
+              "A young orphan, Pip, receives a fortune from a mysterious benefactor and travels to London in order to become a gentleman",
+            industryIdentifiers: [
+              {
+                type: "ISBN_13",
+                identifier: "9780141439563",
+              },
+              {
+                type: "ISBN_10",
+                identifier: "0141439564",
+              },
+            ],
+            readingModes: {
+              text: true,
+              image: false,
+            },
+            pageCount: 514,
+            printType: "BOOK",
+            categories: ["Fiction"],
+            averageRating: 5,
+            ratingsCount: 1,
+            maturityRating: "NOT_MATURE",
+            allowAnonLogging: false,
+            contentVersion: "0.4.2.0.preview.2",
+            panelizationSummary: {
+              containsEpubBubbles: false,
+              containsImageBubbles: false,
+            },
+            imageLinks: {
+              smallThumbnail:
+                "http://books.google.com/books/content?id=KJvuJnzbtJAC&printsec=frontcover&img=1&zoom=5&edge=curl&source=gbs_api",
+              thumbnail:
+                "http://books.google.com/books/content?id=KJvuJnzbtJAC&printsec=frontcover&img=1&zoom=1&edge=curl&source=gbs_api",
+            },
+            language: "en",
+            previewLink:
+              "http://books.google.co.uk/books?id=KJvuJnzbtJAC&printsec=frontcover&dq=isbn:9780141439563&hl=&cd=1&source=gbs_api",
+            infoLink:
+              "http://books.google.co.uk/books?id=KJvuJnzbtJAC&dq=isbn:9780141439563&hl=&source=gbs_api",
+            canonicalVolumeLink:
+              "https://books.google.com/books/about/Great_Expectations.html?hl=&id=KJvuJnzbtJAC",
+          },
+          saleInfo: {
+            country: "GB",
+            saleability: "NOT_FOR_SALE",
+            isEbook: false,
+          },
+          accessInfo: {
+            country: "GB",
+            viewability: "PARTIAL",
+            embeddable: true,
+            publicDomain: false,
+            textToSpeechPermission: "ALLOWED",
+            epub: {
+              isAvailable: true,
+              acsTokenLink:
+                "http://books.google.co.uk/books/download/Great_Expectations-sample-epub.acsm?id=KJvuJnzbtJAC&format=epub&output=acs4_fulfillment_token&dl_type=sample&source=gbs_api",
+            },
+            pdf: {
+              isAvailable: false,
+            },
+            webReaderLink:
+              "http://play.google.com/books/reader?id=KJvuJnzbtJAC&hl=&printsec=frontcover&source=gbs_api",
+            accessViewStatus: "SAMPLE",
+            quoteSharingAllowed: false,
+          },
+          searchInfo: {
+            textSnippet:
+              "Dickens&#39;s haunting late novel depicts Pip&#39;s education and development through adversity as he discovers the true nature of his &#39;great expectations&#39;. This definitive edition uses the text from the first published edition of 1861.",
+          },
+        },
+      ],
+    },
+    {
+      kind: "books#volumes",
+      totalItems: 1,
+      items: [
+        {
+          kind: "books#volume",
+          id: "iaD7DAAAQBAJ",
+          etag: "HNssZO4m7VU",
+          selfLink: "https://www.googleapis.com/books/v1/volumes/iaD7DAAAQBAJ",
+          volumeInfo: {
+            title: "Charlie and the Chocolate Factory (Colour Edition)",
+            authors: ["Roald Dahl"],
+            publisher: "Penguin UK",
+            publishedDate: "2016-09-13",
+            description:
+              "Roald Dahl's Charlie and the Chocolate Factory in glorious full colour. Mr Willy Wonka is the most extraordinary chocolate maker in the world. And do you know who Charlie is? Charlie Bucket is the hero. The other children in this book are nasty little beasts, called: Augustus Gloop - a great big greedy nincompoop; Veruca Salt - a spoiled brat; Violet Beauregarde - a repulsive little gum-chewer; Mike Teavee - a boy who only watches television. Clutching their Golden Tickets, they arrive at Wonka's chocolate factory. But what mysterious secrets will they discover? Our tour is about to begin. Please don't wander off. Mr Wonka wouldn't like to lose any of you at this stage of the proceedings . . . Look out for new Roald Dahl apps in the App store and Google Play- including the disgusting TWIT OR MISS! inspired by the revolting Twits.",
+            industryIdentifiers: [
+              {
+                type: "ISBN_13",
+                identifier: "9780141378558",
+              },
+              {
+                type: "ISBN_10",
+                identifier: "0141378557",
+              },
+            ],
+            readingModes: {
+              text: true,
+              image: false,
+            },
+            pageCount: 160,
+            printType: "BOOK",
+            categories: ["Juvenile Fiction"],
+            averageRating: 3.5,
+            ratingsCount: 62,
+            maturityRating: "NOT_MATURE",
+            allowAnonLogging: true,
+            contentVersion: "1.5.5.0.preview.2",
+            panelizationSummary: {
+              containsEpubBubbles: false,
+              containsImageBubbles: false,
+            },
+            imageLinks: {
+              smallThumbnail:
+                "http://books.google.com/books/content?id=iaD7DAAAQBAJ&printsec=frontcover&img=1&zoom=5&edge=curl&source=gbs_api",
+              thumbnail:
+                "http://books.google.com/books/content?id=iaD7DAAAQBAJ&printsec=frontcover&img=1&zoom=1&edge=curl&source=gbs_api",
+            },
+            language: "en",
+            previewLink:
+              "http://books.google.co.uk/books?id=iaD7DAAAQBAJ&printsec=frontcover&dq=isbn:9780141378558&hl=&cd=1&source=gbs_api",
+            infoLink:
+              "https://play.google.com/store/books/details?id=iaD7DAAAQBAJ&source=gbs_api",
+            canonicalVolumeLink:
+              "https://play.google.com/store/books/details?id=iaD7DAAAQBAJ",
+          },
+          saleInfo: {
+            country: "GB",
+            saleability: "FOR_SALE",
+            isEbook: true,
+            listPrice: {
+              amount: 4.99,
+              currencyCode: "GBP",
+            },
+            retailPrice: {
+              amount: 4.99,
+              currencyCode: "GBP",
+            },
+            buyLink:
+              "https://play.google.com/store/books/details?id=iaD7DAAAQBAJ&rdid=book-iaD7DAAAQBAJ&rdot=1&source=gbs_api",
+            offers: [
+              {
+                finskyOfferType: 1,
+                listPrice: {
+                  amountInMicros: 4990000,
+                  currencyCode: "GBP",
+                },
+                retailPrice: {
+                  amountInMicros: 4990000,
+                  currencyCode: "GBP",
+                },
+                giftable: true,
+              },
+            ],
+          },
+          accessInfo: {
+            country: "GB",
+            viewability: "PARTIAL",
+            embeddable: true,
+            publicDomain: false,
+            textToSpeechPermission: "ALLOWED",
+            epub: {
+              isAvailable: true,
+              acsTokenLink:
+                "http://books.google.co.uk/books/download/Charlie_and_the_Chocolate_Factory_Colour-sample-epub.acsm?id=iaD7DAAAQBAJ&format=epub&output=acs4_fulfillment_token&dl_type=sample&source=gbs_api",
+            },
+            pdf: {
+              isAvailable: false,
+            },
+            webReaderLink:
+              "http://play.google.com/books/reader?id=iaD7DAAAQBAJ&hl=&printsec=frontcover&source=gbs_api",
+            accessViewStatus: "SAMPLE",
+            quoteSharingAllowed: false,
+          },
+          searchInfo: {
+            textSnippet:
+              "The other children in this book are nasty little beasts, called: Augustus Gloop - a great big greedy nincompoop; Veruca Salt - a spoiled brat; Violet Beauregarde - a repulsive little gum-chewer; Mike Teavee - a boy who only watches ...",
+          },
+        },
+      ],
+    },
+    {
+      kind: "books#volumes",
+      totalItems: 1,
+      items: [
+        {
+          kind: "books#volume",
+          id: "zM-vZ-JiSFYC",
+          etag: "KKyweh1V8Nw",
+          selfLink: "https://www.googleapis.com/books/v1/volumes/zM-vZ-JiSFYC",
+          volumeInfo: {
+            title: "The Hunger Games Trilogy",
+            authors: ["Suzanne Collins"],
+            publisher: "Scholastic Inc.",
+            publishedDate: "2011-05-01",
+            description:
+              "The stunning Hunger Games trilogy is complete! The extraordinary, ground breaking New York Times bestsellers The Hunger Games and Catching Fire, along with the third book in The Hunger Games trilogy by Suzanne Collins, Mockingjay, are available for the first time ever in e-book. Stunning, gripping, and powerful.",
+            industryIdentifiers: [
+              {
+                type: "ISBN_13",
+                identifier: "9780545387200",
+              },
+              {
+                type: "ISBN_10",
+                identifier: "0545387205",
+              },
+            ],
+            readingModes: {
+              text: true,
+              image: true,
+            },
+            pageCount: 1187,
+            printType: "BOOK",
+            categories: ["Young Adult Fiction"],
+            averageRating: 4,
+            ratingsCount: 1484,
+            maturityRating: "NOT_MATURE",
+            allowAnonLogging: true,
+            contentVersion: "1.21.23.0.preview.3",
+            panelizationSummary: {
+              containsEpubBubbles: false,
+              containsImageBubbles: false,
+            },
+            imageLinks: {
+              smallThumbnail:
+                "http://books.google.com/books/content?id=zM-vZ-JiSFYC&printsec=frontcover&img=1&zoom=5&edge=curl&source=gbs_api",
+              thumbnail:
+                "http://books.google.com/books/content?id=zM-vZ-JiSFYC&printsec=frontcover&img=1&zoom=1&edge=curl&source=gbs_api",
+            },
+            language: "en",
+            previewLink:
+              "http://books.google.co.uk/books?id=zM-vZ-JiSFYC&printsec=frontcover&dq=isbn:9780545387200&hl=&cd=1&source=gbs_api",
+            infoLink:
+              "http://books.google.co.uk/books?id=zM-vZ-JiSFYC&dq=isbn:9780545387200&hl=&source=gbs_api",
+            canonicalVolumeLink:
+              "https://books.google.com/books/about/The_Hunger_Games_Trilogy.html?hl=&id=zM-vZ-JiSFYC",
+          },
+          saleInfo: {
+            country: "GB",
+            saleability: "NOT_FOR_SALE",
+            isEbook: false,
+          },
+          accessInfo: {
+            country: "GB",
+            viewability: "PARTIAL",
+            embeddable: true,
+            publicDomain: false,
+            textToSpeechPermission: "ALLOWED",
+            epub: {
+              isAvailable: true,
+              acsTokenLink:
+                "http://books.google.co.uk/books/download/The_Hunger_Games_Trilogy-sample-epub.acsm?id=zM-vZ-JiSFYC&format=epub&output=acs4_fulfillment_token&dl_type=sample&source=gbs_api",
+            },
+            pdf: {
+              isAvailable: true,
+              acsTokenLink:
+                "http://books.google.co.uk/books/download/The_Hunger_Games_Trilogy-sample-pdf.acsm?id=zM-vZ-JiSFYC&format=pdf&output=acs4_fulfillment_token&dl_type=sample&source=gbs_api",
+            },
+            webReaderLink:
+              "http://play.google.com/books/reader?id=zM-vZ-JiSFYC&hl=&printsec=frontcover&source=gbs_api",
+            accessViewStatus: "SAMPLE",
+            quoteSharingAllowed: false,
+          },
+          searchInfo: {
+            textSnippet:
+              "The extraordinary, ground breaking New York Times bestsellers The Hunger Games and Catching Fire, along with the third book in The Hunger Games trilogy by Suzanne Collins, Mockingjay, are available for the first time ever in e-book.",
+          },
+        },
+      ],
+    },
+    {
+      kind: "books#volumes",
+      totalItems: 1,
+      items: [
+        {
+          kind: "books#volume",
+          id: "WZyzDwAAQBAJ",
+          etag: "FQQsFZ5bvwM",
+          selfLink: "https://www.googleapis.com/books/v1/volumes/WZyzDwAAQBAJ",
+          volumeInfo: {
+            title: "About a Boy",
+            authors: ["Nick Hornby"],
+            publisher: "Penguin UK",
+            publishedDate: "2015-07-01",
+            description:
+              "About a Boy is Nick Hornby's comic and heart-warming million-copy bestseller 'How cool was Will Freeman?' Too cool! At thirty-six, he's as hip as a teenager. He's single, child-free, goes to the right clubs and knows which trainers to wear. He's also found a great way to score with women: attend single parents' groups full of available (and grateful) mothers, all hoping to meet a Nice Guy. Which is how Will meets Marcus, the oldest twelve-year-old on the planet. Marcus is a bit strange: he listens to Joni Mitchell and Mozart, looks after his mum and has never owned a pair of trainers. But Marcus latches on to Will - and won't let go. Can Will teach Marcus how to grow up cool? And can Marcus help Will just to grow up? This astonishing novel, now a modern classic, was adapted for the acclaimed 2002 film About A Boy, starring Hugh Grant and Nicholas Hoult. Fans of One Day by David Nicholls and Any Human Heart by William Boyd will devour this book, as will lovers of fiction everywhere.",
+            industryIdentifiers: [
+              {
+                type: "ISBN_13",
+                identifier: "9780241976036",
+              },
+              {
+                type: "ISBN_10",
+                identifier: "0241976030",
+              },
+            ],
+            readingModes: {
+              text: true,
+              image: false,
+            },
+            pageCount: 288,
+            printType: "BOOK",
+            categories: ["Fiction"],
+            averageRating: 3.5,
+            ratingsCount: 65,
+            maturityRating: "NOT_MATURE",
+            allowAnonLogging: false,
+            contentVersion: "1.1.1.0.preview.2",
+            panelizationSummary: {
+              containsEpubBubbles: false,
+              containsImageBubbles: false,
+            },
+            imageLinks: {
+              smallThumbnail:
+                "http://books.google.com/books/content?id=WZyzDwAAQBAJ&printsec=frontcover&img=1&zoom=5&edge=curl&source=gbs_api",
+              thumbnail:
+                "http://books.google.com/books/content?id=WZyzDwAAQBAJ&printsec=frontcover&img=1&zoom=1&edge=curl&source=gbs_api",
+            },
+            language: "en",
+            previewLink:
+              "http://books.google.co.uk/books?id=WZyzDwAAQBAJ&printsec=frontcover&dq=isbn:9780241976036&hl=&cd=1&source=gbs_api",
+            infoLink:
+              "http://books.google.co.uk/books?id=WZyzDwAAQBAJ&dq=isbn:9780241976036&hl=&source=gbs_api",
+            canonicalVolumeLink:
+              "https://books.google.com/books/about/About_a_Boy.html?hl=&id=WZyzDwAAQBAJ",
+          },
+          saleInfo: {
+            country: "GB",
+            saleability: "NOT_FOR_SALE",
+            isEbook: false,
+          },
+          accessInfo: {
+            country: "GB",
+            viewability: "PARTIAL",
+            embeddable: true,
+            publicDomain: false,
+            textToSpeechPermission: "ALLOWED",
+            epub: {
+              isAvailable: true,
+              acsTokenLink:
+                "http://books.google.co.uk/books/download/About_a_Boy-sample-epub.acsm?id=WZyzDwAAQBAJ&format=epub&output=acs4_fulfillment_token&dl_type=sample&source=gbs_api",
+            },
+            pdf: {
+              isAvailable: false,
+            },
+            webReaderLink:
+              "http://play.google.com/books/reader?id=WZyzDwAAQBAJ&hl=&printsec=frontcover&source=gbs_api",
+            accessViewStatus: "SAMPLE",
+            quoteSharingAllowed: false,
+          },
+          searchInfo: {
+            textSnippet:
+              "Can Will teach Marcus how to grow up cool? And can Marcus help Will just to grow up? This astonishing novel, now a modern classic, was adapted for the acclaimed 2002 film About A Boy, starring Hugh Grant and Nicholas Hoult.",
+          },
+        },
+      ],
+    },
+    {
+      kind: "books#volumes",
+      totalItems: 1,
+      items: [
+        {
+          kind: "books#volume",
+          id: "fOrU48-804UC",
+          etag: "mXZQIPFOVWE",
+          selfLink: "https://www.googleapis.com/books/v1/volumes/fOrU48-804UC",
+          volumeInfo: {
+            title: "Men are from Mars, Women are from Venus",
+            subtitle: "The Definitive Guide to Relationships",
+            authors: ["John Gray"],
+            publisher: "HarperCollins UK",
+            publishedDate: "2002",
+            description:
+              "The best-selling relationship book of all time- over seven million copies sold world wide to date- in a freshly designed new format. Learn how to create understanding and communication between the sexes from this classic and authoritative guide.",
+            industryIdentifiers: [
+              {
+                type: "ISBN_13",
+                identifier: "9780007152599",
+              },
+              {
+                type: "ISBN_10",
+                identifier: "0007152590",
+              },
+            ],
+            readingModes: {
+              text: false,
+              image: false,
+            },
+            pageCount: 307,
+            printType: "BOOK",
+            categories: ["Communication in marriage"],
+            averageRating: 3.5,
+            ratingsCount: 30,
+            maturityRating: "NOT_MATURE",
+            allowAnonLogging: false,
+            contentVersion: "0.2.1.0.preview.0",
+            panelizationSummary: {
+              containsEpubBubbles: false,
+              containsImageBubbles: false,
+            },
+            imageLinks: {
+              smallThumbnail:
+                "http://books.google.com/books/content?id=fOrU48-804UC&printsec=frontcover&img=1&zoom=5&edge=curl&source=gbs_api",
+              thumbnail:
+                "http://books.google.com/books/content?id=fOrU48-804UC&printsec=frontcover&img=1&zoom=1&edge=curl&source=gbs_api",
+            },
+            language: "en",
+            previewLink:
+              "http://books.google.co.uk/books?id=fOrU48-804UC&printsec=frontcover&dq=isbn:9780007152599&hl=&cd=1&source=gbs_api",
+            infoLink:
+              "http://books.google.co.uk/books?id=fOrU48-804UC&dq=isbn:9780007152599&hl=&source=gbs_api",
+            canonicalVolumeLink:
+              "https://books.google.com/books/about/Men_are_from_Mars_Women_are_from_Venus.html?hl=&id=fOrU48-804UC",
+          },
+          saleInfo: {
+            country: "GB",
+            saleability: "NOT_FOR_SALE",
+            isEbook: false,
+          },
+          accessInfo: {
+            country: "GB",
+            viewability: "PARTIAL",
+            embeddable: false,
+            publicDomain: false,
+            textToSpeechPermission: "ALLOWED_FOR_ACCESSIBILITY",
+            epub: {
+              isAvailable: false,
+            },
+            pdf: {
+              isAvailable: false,
+            },
+            webReaderLink:
+              "http://play.google.com/books/reader?id=fOrU48-804UC&hl=&printsec=frontcover&source=gbs_api",
+            accessViewStatus: "SAMPLE",
+            quoteSharingAllowed: false,
+          },
+          searchInfo: {
+            textSnippet:
+              "The best-selling relationship book of all time- over seven million copies sold world wide to date- in a freshly designed new format.",
+          },
+        },
+      ],
+    },
+    {
+      kind: "books#volumes",
+      totalItems: 1,
+      items: [
+        {
+          kind: "books#volume",
+          id: "yNMdhHubFK0C",
+          etag: "mI+wa5TMEpI",
+          selfLink: "https://www.googleapis.com/books/v1/volumes/yNMdhHubFK0C",
+          volumeInfo: {
+            title: "The Boy in the Striped Pyjamas",
+            authors: ["John Boyne"],
+            publisher: "Random House",
+            publishedDate: "2012-08-02",
+            description:
+              "Discover an extraordinary tale of innocence, friendship and the horrors of war. 'Some things are just sitting there, minding their own business, waiting to be discovered. Like America. And other things are probably better off left alone' Nine-year-old Bruno has a lot of things on his mind. Who is the 'Fury'? Why did he make them leave their nice home in Berlin to go to 'Out-With' ? And who are all the sad people in striped pyjamas on the other side of the fence? The grown-ups won't explain so Bruno decides there is only one thing for it - he will have to explore this place alone. What he discovers is a new friend. A boy with the very same birthday. A boy in striped pyjamas. But why can't they ever play together? ‘A small wonder of a book’ Guardian BACKSTORY: Read an interview with the author JOHN BOYNE and learn all about the Second World War in Germany.",
+            industryIdentifiers: [
+              {
+                type: "ISBN_13",
+                identifier: "9781448139880",
+              },
+              {
+                type: "ISBN_10",
+                identifier: "1448139880",
+              },
+            ],
+            readingModes: {
+              text: true,
+              image: false,
+            },
+            pageCount: 256,
+            printType: "BOOK",
+            categories: ["Juvenile Fiction"],
+            averageRating: 4,
+            ratingsCount: 346,
+            maturityRating: "NOT_MATURE",
+            allowAnonLogging: true,
+            contentVersion: "0.12.9.0.preview.2",
+            panelizationSummary: {
+              containsEpubBubbles: false,
+              containsImageBubbles: false,
+            },
+            imageLinks: {
+              smallThumbnail:
+                "http://books.google.com/books/content?id=yNMdhHubFK0C&printsec=frontcover&img=1&zoom=5&edge=curl&source=gbs_api",
+              thumbnail:
+                "http://books.google.com/books/content?id=yNMdhHubFK0C&printsec=frontcover&img=1&zoom=1&edge=curl&source=gbs_api",
+            },
+            language: "en",
+            previewLink:
+              "http://books.google.co.uk/books?id=yNMdhHubFK0C&printsec=frontcover&dq=isbn:9781448139880&hl=&cd=1&source=gbs_api",
+            infoLink:
+              "https://play.google.com/store/books/details?id=yNMdhHubFK0C&source=gbs_api",
+            canonicalVolumeLink:
+              "https://play.google.com/store/books/details?id=yNMdhHubFK0C",
+          },
+          saleInfo: {
+            country: "GB",
+            saleability: "FOR_SALE",
+            isEbook: true,
+            listPrice: {
+              amount: 2.99,
+              currencyCode: "GBP",
+            },
+            retailPrice: {
+              amount: 2.99,
+              currencyCode: "GBP",
+            },
+            buyLink:
+              "https://play.google.com/store/books/details?id=yNMdhHubFK0C&rdid=book-yNMdhHubFK0C&rdot=1&source=gbs_api",
+            offers: [
+              {
+                finskyOfferType: 1,
+                listPrice: {
+                  amountInMicros: 2990000,
+                  currencyCode: "GBP",
+                },
+                retailPrice: {
+                  amountInMicros: 2990000,
+                  currencyCode: "GBP",
+                },
+                giftable: true,
+              },
+            ],
+          },
+          accessInfo: {
+            country: "GB",
+            viewability: "PARTIAL",
+            embeddable: true,
+            publicDomain: false,
+            textToSpeechPermission: "ALLOWED",
+            epub: {
+              isAvailable: true,
+              acsTokenLink:
+                "http://books.google.co.uk/books/download/The_Boy_in_the_Striped_Pyjamas-sample-epub.acsm?id=yNMdhHubFK0C&format=epub&output=acs4_fulfillment_token&dl_type=sample&source=gbs_api",
+            },
+            pdf: {
+              isAvailable: false,
+            },
+            webReaderLink:
+              "http://play.google.com/books/reader?id=yNMdhHubFK0C&hl=&printsec=frontcover&source=gbs_api",
+            accessViewStatus: "SAMPLE",
+            quoteSharingAllowed: false,
+          },
+          searchInfo: {
+            textSnippet:
+              "A boy in striped pyjamas. But why can&#39;t they ever play together? ‘A small wonder of a book’ Guardian BACKSTORY: Read an interview with the author JOHN BOYNE and learn all about the Second World War in Germany.",
+          },
+        },
+      ],
+    },
+    {
+      kind: "books#volumes",
+      totalItems: 1,
+      items: [
+        {
+          kind: "books#volume",
+          id: "aFACDAAAQBAJ",
+          etag: "2PhcxiZYZNg",
+          selfLink: "https://www.googleapis.com/books/v1/volumes/aFACDAAAQBAJ",
+          volumeInfo: {
+            title: "The Help",
+            authors: ["Kathryn Stockett"],
+            publisher: "Penguin",
+            publishedDate: "2011",
+            description:
+              "Limited and persecuted by racial divides in 1962 Jackson, Mississippi, three women, including an African-American maid, her sassy and chronically unemployed friend and a recently graduated white woman, team up for a clandestine project against a backdrop of the budding civil rights era. Includes reading-group guide. Reissue. A #1 best-selling novel.",
+            industryIdentifiers: [
+              {
+                type: "ISBN_13",
+                identifier: "9780425245132",
+              },
+              {
+                type: "ISBN_10",
+                identifier: "0425245136",
+              },
+            ],
+            readingModes: {
+              text: false,
+              image: false,
+            },
+            pageCount: 534,
+            printType: "BOOK",
+            categories: ["Fiction"],
+            averageRating: 4,
+            ratingsCount: 3273,
+            maturityRating: "NOT_MATURE",
+            allowAnonLogging: false,
+            contentVersion: "preview-1.0.0",
+            panelizationSummary: {
+              containsEpubBubbles: false,
+              containsImageBubbles: false,
+            },
+            imageLinks: {
+              smallThumbnail:
+                "http://books.google.com/books/content?id=aFACDAAAQBAJ&printsec=frontcover&img=1&zoom=5&source=gbs_api",
+              thumbnail:
+                "http://books.google.com/books/content?id=aFACDAAAQBAJ&printsec=frontcover&img=1&zoom=1&source=gbs_api",
+            },
+            language: "en",
+            previewLink:
+              "http://books.google.co.uk/books?id=aFACDAAAQBAJ&dq=isbn:9780425245132&hl=&cd=1&source=gbs_api",
+            infoLink:
+              "http://books.google.co.uk/books?id=aFACDAAAQBAJ&dq=isbn:9780425245132&hl=&source=gbs_api",
+            canonicalVolumeLink:
+              "https://books.google.com/books/about/The_Help.html?hl=&id=aFACDAAAQBAJ",
+          },
+          saleInfo: {
+            country: "GB",
+            saleability: "NOT_FOR_SALE",
+            isEbook: false,
+          },
+          accessInfo: {
+            country: "GB",
+            viewability: "NO_PAGES",
+            embeddable: false,
+            publicDomain: false,
+            textToSpeechPermission: "ALLOWED",
+            epub: {
+              isAvailable: false,
+            },
+            pdf: {
+              isAvailable: false,
+            },
+            webReaderLink:
+              "http://play.google.com/books/reader?id=aFACDAAAQBAJ&hl=&printsec=frontcover&source=gbs_api",
+            accessViewStatus: "NONE",
+            quoteSharingAllowed: false,
+          },
+          searchInfo: {
+            textSnippet:
+              "Limited and persecuted by racial divides in 1962 Jackson, Mississippi, three women, including an African-American maid, her sassy and chronically unemployed friend and a recently graduated white woman, team up for a clandestine project ...",
+          },
+        },
+      ],
+    },
+    {
+      kind: "books#volumes",
+      totalItems: 1,
+      items: [
+        {
+          kind: "books#volume",
+          id: "-aiAOfD8uWAC",
+          etag: "WNCRngPv49I",
+          selfLink: "https://www.googleapis.com/books/v1/volumes/-aiAOfD8uWAC",
+          volumeInfo: {
+            title: "Life of Pi",
+            authors: ["Yann Martel"],
+            publisher: "Canongate Books",
+            publishedDate: "2002",
+            description:
+              "The dazzling, thought-provoking novel that won the Man Booker Prize and became an international phenomenon.",
+            industryIdentifiers: [
+              {
+                type: "ISBN_13",
+                identifier: "9781847676016",
+              },
+              {
+                type: "ISBN_10",
+                identifier: "1847676014",
+              },
+            ],
+            readingModes: {
+              text: false,
+              image: true,
+            },
+            pageCount: 325,
+            printType: "BOOK",
+            categories: ["Shipwreck victims"],
+            averageRating: 3.5,
+            ratingsCount: 3298,
+            maturityRating: "NOT_MATURE",
+            allowAnonLogging: false,
+            contentVersion: "0.0.1.0.preview.1",
+            imageLinks: {
+              smallThumbnail:
+                "http://books.google.com/books/content?id=-aiAOfD8uWAC&printsec=frontcover&img=1&zoom=5&edge=curl&source=gbs_api",
+              thumbnail:
+                "http://books.google.com/books/content?id=-aiAOfD8uWAC&printsec=frontcover&img=1&zoom=1&edge=curl&source=gbs_api",
+            },
+            language: "en",
+            previewLink:
+              "http://books.google.co.uk/books?id=-aiAOfD8uWAC&printsec=frontcover&dq=isbn:9781847676016&hl=&cd=1&source=gbs_api",
+            infoLink:
+              "http://books.google.co.uk/books?id=-aiAOfD8uWAC&dq=isbn:9781847676016&hl=&source=gbs_api",
+            canonicalVolumeLink:
+              "https://books.google.com/books/about/Life_of_Pi.html?hl=&id=-aiAOfD8uWAC",
+          },
+          saleInfo: {
+            country: "GB",
+            saleability: "NOT_FOR_SALE",
+            isEbook: false,
+          },
+          accessInfo: {
+            country: "GB",
+            viewability: "PARTIAL",
+            embeddable: true,
+            publicDomain: false,
+            textToSpeechPermission: "ALLOWED",
+            epub: {
+              isAvailable: false,
+            },
+            pdf: {
+              isAvailable: false,
+            },
+            webReaderLink:
+              "http://play.google.com/books/reader?id=-aiAOfD8uWAC&hl=&printsec=frontcover&source=gbs_api",
+            accessViewStatus: "SAMPLE",
+            quoteSharingAllowed: false,
+          },
+          searchInfo: {
+            textSnippet:
+              "Possessing encyclopedia-like intelligence, unusual zookeeper&#39;s son Pi Patel sets sail for America, but when the ship sinks, he escapes on a life boat and is lost at sea with a dwindling number of animals until only he and a hungry Bengal ...",
+          },
+        },
+      ],
+    },
+    {
+      kind: "books#volumes",
+      totalItems: 1,
+      items: [
+        {
+          kind: "books#volume",
+          id: "JEvFDAAAQBAJ",
+          etag: "l2RhzPgPFCo",
+          selfLink: "https://www.googleapis.com/books/v1/volumes/JEvFDAAAQBAJ",
+          volumeInfo: {
+            title: "The Gruffalo's Child",
+            authors: ["Julia Donaldson"],
+            publisher: "Pan Macmillan",
+            publishedDate: "2016-04-21",
+            description:
+              "\"The Gruffalo said that no gruffalo shouldEver set foot in the deep dark wood.\"But one wild and windy night the Gruffalo's child disobeys her father's warnings and ventures out into the snow. After all, the Big Bad Mouse doesn't really exist . . . does he?The Gruffalo's Child is the Number One best-selling, much-loved sequel to the worldwide picture-book phenomenon that is The Gruffalo. Julia Donaldson's trademark rhyming text and Axel Scheffler's brilliant, characterful illustrations combine once more to ensure that the Gruffalo's Child has followed firmly in her father's footsteps and that her story is one that children will ask for again, and again... and again! This edition features the classic story with a stunning redesigned cover and beautiful finish, making it a must-have addition to the bookshelves of all Donaldson and Scheffler fans - big and small!Also available with redesigned covers are The Gruffalo, Room on the Broom, The Snail and the Whale, The Smartest Giant in Town, Monkey Puzzle, Charlie Cook's Favourite Book, and A Squash and a Squeeze.",
+            industryIdentifiers: [
+              {
+                type: "ISBN_13",
+                identifier: "9781509804764",
+              },
+              {
+                type: "ISBN_10",
+                identifier: "1509804765",
+              },
+            ],
+            readingModes: {
+              text: false,
+              image: false,
+            },
+            pageCount: 32,
+            printType: "BOOK",
+            averageRating: 3.5,
+            ratingsCount: 9,
+            maturityRating: "NOT_MATURE",
+            allowAnonLogging: false,
+            contentVersion: "0.1.0.0.preview.0",
+            panelizationSummary: {
+              containsEpubBubbles: false,
+              containsImageBubbles: false,
+            },
+            imageLinks: {
+              smallThumbnail:
+                "http://books.google.com/books/content?id=JEvFDAAAQBAJ&printsec=frontcover&img=1&zoom=5&edge=curl&source=gbs_api",
+              thumbnail:
+                "http://books.google.com/books/content?id=JEvFDAAAQBAJ&printsec=frontcover&img=1&zoom=1&edge=curl&source=gbs_api",
+            },
+            language: "en",
+            previewLink:
+              "http://books.google.co.uk/books?id=JEvFDAAAQBAJ&printsec=frontcover&dq=isbn:9781509804764&hl=&cd=1&source=gbs_api",
+            infoLink:
+              "http://books.google.co.uk/books?id=JEvFDAAAQBAJ&dq=isbn:9781509804764&hl=&source=gbs_api",
+            canonicalVolumeLink:
+              "https://books.google.com/books/about/The_Gruffalo_s_Child.html?hl=&id=JEvFDAAAQBAJ",
+          },
+          saleInfo: {
+            country: "GB",
+            saleability: "NOT_FOR_SALE",
+            isEbook: false,
+          },
+          accessInfo: {
+            country: "GB",
+            viewability: "PARTIAL",
+            embeddable: true,
+            publicDomain: false,
+            textToSpeechPermission: "ALLOWED_FOR_ACCESSIBILITY",
+            epub: {
+              isAvailable: false,
+            },
+            pdf: {
+              isAvailable: false,
+            },
+            webReaderLink:
+              "http://play.google.com/books/reader?id=JEvFDAAAQBAJ&hl=&printsec=frontcover&source=gbs_api",
+            accessViewStatus: "SAMPLE",
+            quoteSharingAllowed: false,
+          },
+          searchInfo: {
+            textSnippet:
+              "The Gruffalo said that no gruffalo should ever set foot in the deep dark wood ...",
+          },
+        },
+      ],
+    },
+    {
+      kind: "books#volumes",
+      totalItems: 1,
+      items: [
+        {
+          kind: "books#volume",
+          id: "So4NOO9Yq54C",
+          etag: "CDKOhwsnPSg",
+          selfLink: "https://www.googleapis.com/books/v1/volumes/So4NOO9Yq54C",
+          volumeInfo: {
+            title: "The Da Vinci Code",
+            authors: ["Dan Brown"],
+            publisher: "Random House",
+            publishedDate: "2013",
+            description:
+              "The stakes are raised when Langdon uncovers a startling link. The late curator was involved in the Priory of Sion - an actual secret society whose members included Sir Isaac Newton, Botticelli, Victor Hugo, and Da Vinci, among others. Langdon suspects they are on the hunt for a breathtaking historical secret, one that has proven through the centuries to be as enlightening as it is dangerous. In a frantic race through Paris, and beyond, Langdon and Neveu find themselves matching wits with a faceless powerbroker who appears to anticipate their every move. Unless they can decipher the labyrinthine puzzle, the Priory's secret - and an explosive ancient truth - will be lost forever.",
+            industryIdentifiers: [
+              {
+                type: "ISBN_13",
+                identifier: "9780552169929",
+              },
+              {
+                type: "ISBN_10",
+                identifier: "0552169927",
+              },
+            ],
+            readingModes: {
+              text: false,
+              image: false,
+            },
+            pageCount: 606,
+            printType: "BOOK",
+            categories: ["Art museum curators"],
+            averageRating: 3,
+            ratingsCount: 3450,
+            maturityRating: "NOT_MATURE",
+            allowAnonLogging: false,
+            contentVersion: "0.6.0.0.preview.0",
+            panelizationSummary: {
+              containsEpubBubbles: false,
+              containsImageBubbles: false,
+            },
+            imageLinks: {
+              smallThumbnail:
+                "http://books.google.com/books/content?id=So4NOO9Yq54C&printsec=frontcover&img=1&zoom=5&edge=curl&source=gbs_api",
+              thumbnail:
+                "http://books.google.com/books/content?id=So4NOO9Yq54C&printsec=frontcover&img=1&zoom=1&edge=curl&source=gbs_api",
+            },
+            language: "en",
+            previewLink:
+              "http://books.google.co.uk/books?id=So4NOO9Yq54C&printsec=frontcover&dq=isbn:9780552169929&hl=&cd=1&source=gbs_api",
+            infoLink:
+              "http://books.google.co.uk/books?id=So4NOO9Yq54C&dq=isbn:9780552169929&hl=&source=gbs_api",
+            canonicalVolumeLink:
+              "https://books.google.com/books/about/The_Da_Vinci_Code.html?hl=&id=So4NOO9Yq54C",
+          },
+          saleInfo: {
+            country: "GB",
+            saleability: "NOT_FOR_SALE",
+            isEbook: false,
+          },
+          accessInfo: {
+            country: "GB",
+            viewability: "PARTIAL",
+            embeddable: true,
+            publicDomain: false,
+            textToSpeechPermission: "ALLOWED",
+            epub: {
+              isAvailable: false,
+            },
+            pdf: {
+              isAvailable: false,
+            },
+            webReaderLink:
+              "http://play.google.com/books/reader?id=So4NOO9Yq54C&hl=&printsec=frontcover&source=gbs_api",
+            accessViewStatus: "SAMPLE",
+            quoteSharingAllowed: false,
+          },
+          searchInfo: {
+            textSnippet:
+              "As he and a gifted French cryptologist, Sophie Neveu, sort through the bizarre riddles, they are stunned to find a trail that leads to the works of Leonardo Da Vinci - and suggests the answer to an age-old mystery that stretches deep into ...",
+          },
+        },
+      ],
+    },
+    {
+      kind: "books#volumes",
+      totalItems: 1,
+      items: [
+        {
+          kind: "books#volume",
+          id: "cH57BtSHMkQC",
+          etag: "EGOIw62jmbA",
+          selfLink: "https://www.googleapis.com/books/v1/volumes/cH57BtSHMkQC",
+          volumeInfo: {
+            title: "The Dalai Lama's Book of Wisdom",
+            authors: ["Dalai Lama XIV Bstan-ʼdzin-rgya-mtsho", "Dalai Lama"],
+            publisher: "HarperCollins UK",
+            publishedDate: "1999",
+            description:
+              "In this text, His Holiness the Dalai Lama describes how to bring wisdom and compassion into our busy, stressful everyday lives. A selection of words from His Holiness aims to help the reader to face difficult emotions such as anger in themselves and in others with acceptance and understanding.",
+            industryIdentifiers: [
+              {
+                type: "ISBN_13",
+                identifier: "9780722539552",
+              },
+              {
+                type: "ISBN_10",
+                identifier: "072253955X",
+              },
+            ],
+            readingModes: {
+              text: false,
+              image: false,
+            },
+            pageCount: 115,
+            printType: "BOOK",
+            categories: ["Religion"],
+            averageRating: 5,
+            ratingsCount: 1,
+            maturityRating: "NOT_MATURE",
+            allowAnonLogging: false,
+            contentVersion: "0.2.3.0.preview.0",
+            panelizationSummary: {
+              containsEpubBubbles: false,
+              containsImageBubbles: false,
+            },
+            imageLinks: {
+              smallThumbnail:
+                "http://books.google.com/books/content?id=cH57BtSHMkQC&printsec=frontcover&img=1&zoom=5&edge=curl&source=gbs_api",
+              thumbnail:
+                "http://books.google.com/books/content?id=cH57BtSHMkQC&printsec=frontcover&img=1&zoom=1&edge=curl&source=gbs_api",
+            },
+            language: "en",
+            previewLink:
+              "http://books.google.co.uk/books?id=cH57BtSHMkQC&printsec=frontcover&dq=isbn:9780722539552&hl=&cd=1&source=gbs_api",
+            infoLink:
+              "http://books.google.co.uk/books?id=cH57BtSHMkQC&dq=isbn:9780722539552&hl=&source=gbs_api",
+            canonicalVolumeLink:
+              "https://books.google.com/books/about/The_Dalai_Lama_s_Book_of_Wisdom.html?hl=&id=cH57BtSHMkQC",
+          },
+          saleInfo: {
+            country: "GB",
+            saleability: "NOT_FOR_SALE",
+            isEbook: false,
+          },
+          accessInfo: {
+            country: "GB",
+            viewability: "PARTIAL",
+            embeddable: false,
+            publicDomain: false,
+            textToSpeechPermission: "ALLOWED_FOR_ACCESSIBILITY",
+            epub: {
+              isAvailable: false,
+            },
+            pdf: {
+              isAvailable: false,
+            },
+            webReaderLink:
+              "http://play.google.com/books/reader?id=cH57BtSHMkQC&hl=&printsec=frontcover&source=gbs_api",
+            accessViewStatus: "SAMPLE",
+            quoteSharingAllowed: false,
+          },
+          searchInfo: {
+            textSnippet:
+              "In this text, His Holiness the Dalai Lama describes how to bring wisdom and compassion into our busy, stressful everyday lives.",
+          },
+        },
+      ],
+    },
+    {
+      kind: "books#volumes",
+      totalItems: 1,
+      items: [
+        {
+          kind: "books#volume",
+          id: "PjgjrgEACAAJ",
+          etag: "EEK3JkDDIzo",
+          selfLink: "https://www.googleapis.com/books/v1/volumes/PjgjrgEACAAJ",
+          volumeInfo: {
+            title: "In Search of Lost Time",
+            authors: [
+              "Marcel Proust",
+              "C. K. Scott Moncrieff",
+              "Vincent Kelvin",
+            ],
+            publisher: "CreateSpace",
+            publishedDate: "2015-01-31",
+            description:
+              'In Search of Lost Time (French: À la recherche du temps perdu)—also translated as Remembrance of Things Past—is a novel by Marcel Proust (1871–1922). His most prominent work, it is known both for its length and its theme of involuntary memory, the most famous example being the "episode of the madeleine." The novel began to take shape in 1909. Proust continued to work on it until his final illness in the autumn of 1922 forced him to break off. Proust established the structure early on, but even after volumes were initially finished he kept adding new material and edited one volume after another for publication. The last three of the seven volumes contain oversights and fragmentary or unpolished passages as they existed in draft form at the death of the author; the publication of these parts was overseen by his brother Robert.The work was published in France between 1913 and 1927. Proust paid for the publication of the first volume (by the Grasset publishing house) after it had been turned down by leading editors who had been offered the manuscript in longhand. Many of its ideas, motifs, and scenes are foreshadowed in Proust\'s unfinished novel, Jean Santeuil (1896–99), though the perspective and treatment there are different, and in his unfinished hybrid of philosophical essay and story, Contre Sainte-Beuve (1908–09). The novel had great influence on twentieth-century literature; some writers have sought to emulate it, others to parody it. In the centenary year of Du côté de chez Swann, Edmund White pronounced A la recherche du temps perdu "the most respected novel of the twentieth century."',
+            industryIdentifiers: [
+              {
+                type: "ISBN_10",
+                identifier: "150778399X",
+              },
+              {
+                type: "ISBN_13",
+                identifier: "9781507783993",
+              },
+            ],
+            readingModes: {
+              text: false,
+              image: false,
+            },
+            pageCount: 660,
+            printType: "BOOK",
+            categories: ["Fiction"],
+            averageRating: 4,
+            ratingsCount: 5,
+            maturityRating: "NOT_MATURE",
+            allowAnonLogging: false,
+            contentVersion: "preview-1.0.0",
+            imageLinks: {
+              smallThumbnail:
+                "http://books.google.com/books/content?id=PjgjrgEACAAJ&printsec=frontcover&img=1&zoom=5&source=gbs_api",
+              thumbnail:
+                "http://books.google.com/books/content?id=PjgjrgEACAAJ&printsec=frontcover&img=1&zoom=1&source=gbs_api",
+            },
+            language: "en",
+            previewLink:
+              "http://books.google.co.uk/books?id=PjgjrgEACAAJ&dq=isbn:9781507783993&hl=&cd=1&source=gbs_api",
+            infoLink:
+              "http://books.google.co.uk/books?id=PjgjrgEACAAJ&dq=isbn:9781507783993&hl=&source=gbs_api",
+            canonicalVolumeLink:
+              "https://books.google.com/books/about/In_Search_of_Lost_Time.html?hl=&id=PjgjrgEACAAJ",
+          },
+          saleInfo: {
+            country: "GB",
+            saleability: "NOT_FOR_SALE",
+            isEbook: false,
+          },
+          accessInfo: {
+            country: "GB",
+            viewability: "NO_PAGES",
+            embeddable: false,
+            publicDomain: false,
+            textToSpeechPermission: "ALLOWED",
+            epub: {
+              isAvailable: false,
+            },
+            pdf: {
+              isAvailable: false,
+            },
+            webReaderLink:
+              "http://play.google.com/books/reader?id=PjgjrgEACAAJ&hl=&printsec=frontcover&source=gbs_api",
+            accessViewStatus: "NONE",
+            quoteSharingAllowed: false,
+          },
+          searchInfo: {
+            textSnippet:
+              "&quot; The novel began to take shape in 1909. Proust continued to work on it until his final illness in the autumn of 1922 forced him to break off.",
+          },
+        },
+      ],
+    },
+    {
+      kind: "books#volumes",
+      totalItems: 1,
+      items: [
+        {
+          kind: "books#volume",
+          id: "9lNoAwAAQBAJ",
+          etag: "3kVhcsA7wUw",
+          selfLink: "https://www.googleapis.com/books/v1/volumes/9lNoAwAAQBAJ",
+          volumeInfo: {
+            title: "Moby Dick",
+            authors: ["Herman Melville"],
+            publisher: "BookRix",
+            publishedDate: "2022-02-01",
+            description:
+              'Moby Dick is a novel by American writer Herman Melville. The work is an epic sea story of Captain Ahab\'s voyage in pursuit of Moby Dick, a great white whale. A contemporary commercial failure and out of print at the time of the author\'s death in 1891, its reputation rose during the twentieth century. D.H. Lawrence called it "the greatest book of the sea ever written." Jorge Luis Borges praised the style: "Unforgettable phrases abound." Today it is considered one of the Great American Novels and a leading work of American Romanticism. The opening line, "Call me Ishmael," is one of the most recognizable opening lines in Western literature. Ishmael then narrates the voyage of the whaleship Pequod, commanded by Captain Ahab. Ahab has one purpose: revenge on Moby Dick, a ferocious, enigmatic white whale which on a previous voyage destroyed Ahab\'s ship and severed his leg at the knee. The detailed and realistic descriptions of whale hunting and the process of extracting whale oil, as well as life aboard ship among a culturally diverse crew, are mixed with exploration of class and social status, good and evil, and the existence of God. Melville uses a wide range of styles and literary devices ranging from lists and catalogs to Shakespearean stage directions, soliloquies, and asides.',
+            industryIdentifiers: [
+              {
+                type: "ISBN_13",
+                identifier: "9783736800748",
+              },
+              {
+                type: "ISBN_10",
+                identifier: "3736800746",
+              },
+            ],
+            readingModes: {
+              text: true,
+              image: true,
+            },
+            pageCount: 834,
+            printType: "BOOK",
+            categories: ["Fiction"],
+            maturityRating: "NOT_MATURE",
+            allowAnonLogging: true,
+            contentVersion: "1.123.117.0.preview.3",
+            panelizationSummary: {
+              containsEpubBubbles: false,
+              containsImageBubbles: false,
+            },
+            imageLinks: {
+              smallThumbnail:
+                "http://books.google.com/books/content?id=9lNoAwAAQBAJ&printsec=frontcover&img=1&zoom=5&edge=curl&source=gbs_api",
+              thumbnail:
+                "http://books.google.com/books/content?id=9lNoAwAAQBAJ&printsec=frontcover&img=1&zoom=1&edge=curl&source=gbs_api",
+            },
+            language: "en",
+            previewLink:
+              "http://books.google.co.uk/books?id=9lNoAwAAQBAJ&printsec=frontcover&dq=isbn:9783736800748&hl=&cd=1&source=gbs_api",
+            infoLink:
+              "https://play.google.com/store/books/details?id=9lNoAwAAQBAJ&source=gbs_api",
+            canonicalVolumeLink:
+              "https://play.google.com/store/books/details?id=9lNoAwAAQBAJ",
+          },
+          saleInfo: {
+            country: "GB",
+            saleability: "FOR_SALE",
+            isEbook: true,
+            listPrice: {
+              amount: 2.99,
+              currencyCode: "GBP",
+            },
+            retailPrice: {
+              amount: 2.99,
+              currencyCode: "GBP",
+            },
+            buyLink:
+              "https://play.google.com/store/books/details?id=9lNoAwAAQBAJ&rdid=book-9lNoAwAAQBAJ&rdot=1&source=gbs_api",
+            offers: [
+              {
+                finskyOfferType: 1,
+                listPrice: {
+                  amountInMicros: 2990000,
+                  currencyCode: "GBP",
+                },
+                retailPrice: {
+                  amountInMicros: 2990000,
+                  currencyCode: "GBP",
+                },
+                giftable: true,
+              },
+            ],
+          },
+          accessInfo: {
+            country: "GB",
+            viewability: "PARTIAL",
+            embeddable: true,
+            publicDomain: false,
+            textToSpeechPermission: "ALLOWED",
+            epub: {
+              isAvailable: true,
+              acsTokenLink:
+                "http://books.google.co.uk/books/download/Moby_Dick-sample-epub.acsm?id=9lNoAwAAQBAJ&format=epub&output=acs4_fulfillment_token&dl_type=sample&source=gbs_api",
+            },
+            pdf: {
+              isAvailable: true,
+              acsTokenLink:
+                "http://books.google.co.uk/books/download/Moby_Dick-sample-pdf.acsm?id=9lNoAwAAQBAJ&format=pdf&output=acs4_fulfillment_token&dl_type=sample&source=gbs_api",
+            },
+            webReaderLink:
+              "http://play.google.com/books/reader?id=9lNoAwAAQBAJ&hl=&printsec=frontcover&source=gbs_api",
+            accessViewStatus: "SAMPLE",
+            quoteSharingAllowed: false,
+          },
+          searchInfo: {
+            textSnippet:
+              "Moby Dick is a novel by American writer Herman Melville. The work is an epic sea story of Captain Ahab&#39;s voyage in pursuit of Moby Dick, a great white whale.",
+          },
+        },
+      ],
+    },
+    {
+      kind: "books#volumes",
+      totalItems: 1,
+      items: [
+        {
+          kind: "books#volume",
+          id: "8lBAdXndV9AC",
+          etag: "8QDy3oAvEW4",
+          selfLink: "https://www.googleapis.com/books/v1/volumes/8lBAdXndV9AC",
+          volumeInfo: {
+            title: "War and Peace",
+            authors: ["graf Leo Tolstoy"],
+            publisher: "Library of Alexandria",
+            publishedDate: "1971",
+            industryIdentifiers: [
+              {
+                type: "ISBN_13",
+                identifier: "9781465510778",
+              },
+              {
+                type: "ISBN_10",
+                identifier: "146551077X",
+              },
+            ],
+            readingModes: {
+              text: true,
+              image: false,
+            },
+            pageCount: 1352,
+            printType: "BOOK",
+            categories: ["English fiction"],
+            averageRating: 3.5,
+            ratingsCount: 11,
+            maturityRating: "NOT_MATURE",
+            allowAnonLogging: true,
+            contentVersion: "2.9.8.0.preview.2",
+            panelizationSummary: {
+              containsEpubBubbles: false,
+              containsImageBubbles: false,
+            },
+            imageLinks: {
+              smallThumbnail:
+                "http://books.google.com/books/content?id=8lBAdXndV9AC&printsec=frontcover&img=1&zoom=5&edge=curl&source=gbs_api",
+              thumbnail:
+                "http://books.google.com/books/content?id=8lBAdXndV9AC&printsec=frontcover&img=1&zoom=1&edge=curl&source=gbs_api",
+            },
+            language: "en",
+            previewLink:
+              "http://books.google.co.uk/books?id=8lBAdXndV9AC&printsec=frontcover&dq=isbn:9781465510778&hl=&cd=1&source=gbs_api",
+            infoLink:
+              "https://play.google.com/store/books/details?id=8lBAdXndV9AC&source=gbs_api",
+            canonicalVolumeLink:
+              "https://play.google.com/store/books/details?id=8lBAdXndV9AC",
+          },
+          saleInfo: {
+            country: "GB",
+            saleability: "FOR_SALE",
+            isEbook: true,
+            listPrice: {
+              amount: 4.99,
+              currencyCode: "GBP",
+            },
+            retailPrice: {
+              amount: 4.99,
+              currencyCode: "GBP",
+            },
+            buyLink:
+              "https://play.google.com/store/books/details?id=8lBAdXndV9AC&rdid=book-8lBAdXndV9AC&rdot=1&source=gbs_api",
+            offers: [
+              {
+                finskyOfferType: 1,
+                listPrice: {
+                  amountInMicros: 4990000,
+                  currencyCode: "GBP",
+                },
+                retailPrice: {
+                  amountInMicros: 4990000,
+                  currencyCode: "GBP",
+                },
+                giftable: true,
+              },
+            ],
+          },
+          accessInfo: {
+            country: "GB",
+            viewability: "PARTIAL",
+            embeddable: true,
+            publicDomain: false,
+            textToSpeechPermission: "ALLOWED",
+            epub: {
+              isAvailable: true,
+              acsTokenLink:
+                "http://books.google.co.uk/books/download/War_and_Peace-sample-epub.acsm?id=8lBAdXndV9AC&format=epub&output=acs4_fulfillment_token&dl_type=sample&source=gbs_api",
+            },
+            pdf: {
+              isAvailable: false,
+            },
+            webReaderLink:
+              "http://play.google.com/books/reader?id=8lBAdXndV9AC&hl=&printsec=frontcover&source=gbs_api",
+            accessViewStatus: "SAMPLE",
+            quoteSharingAllowed: false,
+          },
+        },
+      ],
+    },
+    {
+      kind: "books#volumes",
+      totalItems: 1,
+      items: [
+        {
+          kind: "books#volume",
+          id: "aR0DAwAAQBAJ",
+          etag: "ZwzDiQmgXKM",
+          selfLink: "https://www.googleapis.com/books/v1/volumes/aR0DAwAAQBAJ",
+          volumeInfo: {
+            title: "Hamlet",
+            authors: ["William Shakespeare"],
+            publisher: "Courier Corporation",
+            publishedDate: "1992-09-24",
+            description:
+              "In this quintessential Shakespearean drama, Hamlet's halting pursuit of revenge for his father's death unfolds in a series of highly charged confrontations that climax in tragedy. Probing the depths of human feeling like few other works of art, the play is reprinted here from an authoritative British edition complete with illuminating footnotes. A selection of the Common Core State Standards Initiative.",
+            industryIdentifiers: [
+              {
+                type: "ISBN_13",
+                identifier: "9780486272788",
+              },
+              {
+                type: "ISBN_10",
+                identifier: "0486272788",
+              },
+            ],
+            readingModes: {
+              text: false,
+              image: true,
+            },
+            pageCount: 122,
+            printType: "BOOK",
+            categories: ["Drama"],
+            averageRating: 4.5,
+            ratingsCount: 2,
+            maturityRating: "NOT_MATURE",
+            allowAnonLogging: true,
+            contentVersion: "1.5.1.0.preview.1",
+            panelizationSummary: {
+              containsEpubBubbles: false,
+              containsImageBubbles: false,
+            },
+            imageLinks: {
+              smallThumbnail:
+                "http://books.google.com/books/content?id=aR0DAwAAQBAJ&printsec=frontcover&img=1&zoom=5&edge=curl&source=gbs_api",
+              thumbnail:
+                "http://books.google.com/books/content?id=aR0DAwAAQBAJ&printsec=frontcover&img=1&zoom=1&edge=curl&source=gbs_api",
+            },
+            language: "en",
+            previewLink:
+              "http://books.google.co.uk/books?id=aR0DAwAAQBAJ&printsec=frontcover&dq=isbn:9780486272788&hl=&cd=1&source=gbs_api",
+            infoLink:
+              "http://books.google.co.uk/books?id=aR0DAwAAQBAJ&dq=isbn:9780486272788&hl=&source=gbs_api",
+            canonicalVolumeLink:
+              "https://books.google.com/books/about/Hamlet.html?hl=&id=aR0DAwAAQBAJ",
+          },
+          saleInfo: {
+            country: "GB",
+            saleability: "NOT_FOR_SALE",
+            isEbook: false,
+          },
+          accessInfo: {
+            country: "GB",
+            viewability: "PARTIAL",
+            embeddable: true,
+            publicDomain: false,
+            textToSpeechPermission: "ALLOWED",
+            epub: {
+              isAvailable: false,
+            },
+            pdf: {
+              isAvailable: false,
+            },
+            webReaderLink:
+              "http://play.google.com/books/reader?id=aR0DAwAAQBAJ&hl=&printsec=frontcover&source=gbs_api",
+            accessViewStatus: "SAMPLE",
+            quoteSharingAllowed: false,
+          },
+          searchInfo: {
+            textSnippet:
+              "In this quintessential Shakespearean drama, Hamlet&#39;s halting pursuit of revenge for his father&#39;s death unfolds in a series of highly charged confrontations that climax in tragedy.",
+          },
+        },
+      ],
+    },
+    {
+      kind: "books#volumes",
+      totalItems: 1,
+      items: [
+        {
+          kind: "books#volume",
+          id: "MZAvEAAAQBAJ",
+          etag: "5B7JnmTxgcY",
+          selfLink: "https://www.googleapis.com/books/v1/volumes/MZAvEAAAQBAJ",
+          volumeInfo: {
+            title: "Wuthering Heights",
+            authors: ["Emily Bronte"],
+            publisher: "Diamond Pocket Books Pvt Ltd",
+            publishedDate: "2021-05-21",
+            description:
+              'Emily Bronte was an English novelist & poet, who is best known for her only novel, "Wuthering Heights: She has written poems also such as - ‘Poems by Currer, Ellis and Action Bell’, ‘A Death Scene’, ‘To a Wreath of Snow, and lots Many. ‘Wuthering Heights’ is a highly imaginative work of passion and hate. Author was interested in mysticism and used to enjoy her solitude outdoors. This novel consists of those elements. It is now considered a classic of English literature. It was published under the pseudonym - "Ellis Bell” The story is full of high creativity and very imaginative. It narrates revenge also. It revolves around the main character, Heathcliff. Wuthering Heights is his farmhouse. Heathcliff is a young orphan, who was brought by Earnshaw at Wuthering Heights, 30 years ago. Earnshaw loves him (Heathcliff) so much, even neglects his own children. After death of Earnshaw, his elder son Hindley becomes the new master of Wuthering Heights and he allows Heathcliff to stay there only as a servant. Catherine is in love with Heathcliff, but doesn\'t show due to her social statue. The story thus seems very interesting and it ends with sights of the ghosts of Catherine and Heathcliff. It consists of many ups and downs Readers will Surely going to enjoy the novel. It’s Heartthrobing and it’s very difficult to getup without reading the novel - fully.',
+            industryIdentifiers: [
+              {
+                type: "ISBN_13",
+                identifier: "9789354860805",
+              },
+              {
+                type: "ISBN_10",
+                identifier: "935486080X",
+              },
+            ],
+            readingModes: {
+              text: true,
+              image: false,
+            },
+            pageCount: 272,
+            printType: "BOOK",
+            categories: ["Fiction"],
+            maturityRating: "NOT_MATURE",
+            allowAnonLogging: false,
+            contentVersion: "1.3.3.0.preview.2",
+            panelizationSummary: {
+              containsEpubBubbles: false,
+              containsImageBubbles: false,
+            },
+            imageLinks: {
+              smallThumbnail:
+                "http://books.google.com/books/content?id=MZAvEAAAQBAJ&printsec=frontcover&img=1&zoom=5&edge=curl&source=gbs_api",
+              thumbnail:
+                "http://books.google.com/books/content?id=MZAvEAAAQBAJ&printsec=frontcover&img=1&zoom=1&edge=curl&source=gbs_api",
+            },
+            language: "en",
+            previewLink:
+              "http://books.google.co.uk/books?id=MZAvEAAAQBAJ&printsec=frontcover&dq=isbn:9789354860805&hl=&cd=1&source=gbs_api",
+            infoLink:
+              "https://play.google.com/store/books/details?id=MZAvEAAAQBAJ&source=gbs_api",
+            canonicalVolumeLink:
+              "https://play.google.com/store/books/details?id=MZAvEAAAQBAJ",
+          },
+          saleInfo: {
+            country: "GB",
+            saleability: "FOR_SALE",
+            isEbook: true,
+            listPrice: {
+              amount: 0.29,
+              currencyCode: "GBP",
+            },
+            retailPrice: {
+              amount: 0.29,
+              currencyCode: "GBP",
+            },
+            buyLink:
+              "https://play.google.com/store/books/details?id=MZAvEAAAQBAJ&rdid=book-MZAvEAAAQBAJ&rdot=1&source=gbs_api",
+            offers: [
+              {
+                finskyOfferType: 1,
+                listPrice: {
+                  amountInMicros: 290000,
+                  currencyCode: "GBP",
+                },
+                retailPrice: {
+                  amountInMicros: 290000,
+                  currencyCode: "GBP",
+                },
+                giftable: true,
+              },
+            ],
+          },
+          accessInfo: {
+            country: "GB",
+            viewability: "PARTIAL",
+            embeddable: true,
+            publicDomain: false,
+            textToSpeechPermission: "ALLOWED",
+            epub: {
+              isAvailable: true,
+              acsTokenLink:
+                "http://books.google.co.uk/books/download/Wuthering_Heights-sample-epub.acsm?id=MZAvEAAAQBAJ&format=epub&output=acs4_fulfillment_token&dl_type=sample&source=gbs_api",
+            },
+            pdf: {
+              isAvailable: false,
+            },
+            webReaderLink:
+              "http://play.google.com/books/reader?id=MZAvEAAAQBAJ&hl=&printsec=frontcover&source=gbs_api",
+            accessViewStatus: "SAMPLE",
+            quoteSharingAllowed: false,
+          },
+          searchInfo: {
+            textSnippet:
+              "This novel consists of those elements. It is now considered a classic of English literature. It was published under the pseudonym - &quot;Ellis Bell” The story is full of high creativity and very imaginative. It narrates revenge also.",
+          },
+        },
+      ],
+    },
+    {
+      kind: "books#volumes",
+      totalItems: 1,
+      items: [
+        {
+          kind: "books#volume",
+          id: "ZxwpakTv_MIC",
+          etag: "Nr+Qw6UA8Rk",
+          selfLink: "https://www.googleapis.com/books/v1/volumes/ZxwpakTv_MIC",
+          volumeInfo: {
+            title: "Alice in Wonderland",
+            authors: ["Lewis Carroll"],
+            publisher: "The Floating Press",
+            publishedDate: "2009-01-01",
+            description:
+              "Alice in Wonderland (also known as Alice's Adventures in Wonderland), from 1865, is the peculiar and imaginative tale of a girl who falls down a rabbit-hole into a bizarre world of eccentric and unusual creatures. Lewis Carroll's prominent example of the genre of \"literary nonsense\" has endured in popularity with its clever way of playing with logic and a narrative structure that has influence generations of fiction writing.",
+            industryIdentifiers: [
+              {
+                type: "ISBN_13",
+                identifier: "9781877527814",
+              },
+              {
+                type: "ISBN_10",
+                identifier: "1877527815",
+              },
+            ],
+            readingModes: {
+              text: true,
+              image: false,
+            },
+            pageCount: 104,
+            printType: "BOOK",
+            categories: ["Fiction"],
+            averageRating: 3.5,
+            ratingsCount: 177,
+            maturityRating: "NOT_MATURE",
+            allowAnonLogging: true,
+            contentVersion: "0.2.3.0.preview.2",
+            panelizationSummary: {
+              containsEpubBubbles: false,
+              containsImageBubbles: false,
+            },
+            imageLinks: {
+              smallThumbnail:
+                "http://books.google.com/books/content?id=ZxwpakTv_MIC&printsec=frontcover&img=1&zoom=5&edge=curl&source=gbs_api",
+              thumbnail:
+                "http://books.google.com/books/content?id=ZxwpakTv_MIC&printsec=frontcover&img=1&zoom=1&edge=curl&source=gbs_api",
+            },
+            language: "en",
+            previewLink:
+              "http://books.google.co.uk/books?id=ZxwpakTv_MIC&printsec=frontcover&dq=isbn:9781877527814&hl=&cd=1&source=gbs_api",
+            infoLink:
+              "https://play.google.com/store/books/details?id=ZxwpakTv_MIC&source=gbs_api",
+            canonicalVolumeLink:
+              "https://play.google.com/store/books/details?id=ZxwpakTv_MIC",
+          },
+          saleInfo: {
+            country: "GB",
+            saleability: "FOR_SALE",
+            isEbook: true,
+            listPrice: {
+              amount: 2.69,
+              currencyCode: "GBP",
+            },
+            retailPrice: {
+              amount: 2.69,
+              currencyCode: "GBP",
+            },
+            buyLink:
+              "https://play.google.com/store/books/details?id=ZxwpakTv_MIC&rdid=book-ZxwpakTv_MIC&rdot=1&source=gbs_api",
+            offers: [
+              {
+                finskyOfferType: 1,
+                listPrice: {
+                  amountInMicros: 2690000,
+                  currencyCode: "GBP",
+                },
+                retailPrice: {
+                  amountInMicros: 2690000,
+                  currencyCode: "GBP",
+                },
+                giftable: true,
+              },
+            ],
+          },
+          accessInfo: {
+            country: "GB",
+            viewability: "PARTIAL",
+            embeddable: true,
+            publicDomain: false,
+            textToSpeechPermission: "ALLOWED",
+            epub: {
+              isAvailable: true,
+              acsTokenLink:
+                "http://books.google.co.uk/books/download/Alice_in_Wonderland-sample-epub.acsm?id=ZxwpakTv_MIC&format=epub&output=acs4_fulfillment_token&dl_type=sample&source=gbs_api",
+            },
+            pdf: {
+              isAvailable: false,
+            },
+            webReaderLink:
+              "http://play.google.com/books/reader?id=ZxwpakTv_MIC&hl=&printsec=frontcover&source=gbs_api",
+            accessViewStatus: "SAMPLE",
+            quoteSharingAllowed: false,
+          },
+          searchInfo: {
+            textSnippet:
+              "Alice in Wonderland (also known as Alice&#39;s Adventures in Wonderland), from 1865, is the peculiar and imaginative tale of a girl who falls down a rabbit-hole into a bizarre world of eccentric and unusual creatures.",
+          },
+        },
+      ],
+    },
+    {
+      kind: "books#volumes",
+      totalItems: 1,
+      items: [
+        {
+          kind: "books#volume",
+          id: "hqsJKvM17WwC",
+          etag: "yzkcmBu2Dhk",
+          selfLink: "https://www.googleapis.com/books/v1/volumes/hqsJKvM17WwC",
+          volumeInfo: {
+            title: "Gulliver's Travels",
+            authors: ["Jonathan Swift"],
+            publisher: "Penguin",
+            publishedDate: "2010-11-02",
+            description:
+              "\"It is universally read, from the cabinet council to the nursery,\" remarked Alexander Pope when Gulliver's Travels was published in 1726. One of the unique books of world literature, Swift's masterful satire describes the astonishing voyages of one Lemuel Gulliver, a ship's surgeon, to surreal kingdoms inhabited by miniature people and giants, quack philosophers and scientists, horses endowed with reason and men who behave like beasts. Written with great wit and invention, Gulliver's Travels is a savage parody on man and his institutions that has captivated readers for nearly three centuries. The Modern Library has played a significant role in American cultural life for the better part of a century. The series was founded in 1917 by the publishers Boni and Liveright and eight years later acquired by Bennett Cerf and Donald Klopfer. It provided the foundation for their next publishing venture, Random House. The Modern Library has been a staple of the American book trade, providing readers with afford- able hardbound editions of impor- tant works of literature and thought. For the Modern Library's seventy- fifth anniversary, Random House redesigned the series, restoring as its emblem the running torch- bearer created by Lucian Bernhard in 1925 and refurbishing jackets, bindings, and type, as well as inau- gurating a new program of selecting titles. The Modern Library continues to provide the world's best books, at the best prices. As bestselling author and critic Allan Bloom observed: \"Gulliver's Travels is an amazing rhetorical achievement. Swift had not only the judgment with which to arrive at a reasoned view of the world but the fancy by means of which he could re-create that world in a form which teaches where argument fails and which satisfies all while misleading none.\"",
+            industryIdentifiers: [
+              {
+                type: "ISBN_13",
+                identifier: "9781101477816",
+              },
+              {
+                type: "ISBN_10",
+                identifier: "1101477814",
+              },
+            ],
+            readingModes: {
+              text: true,
+              image: false,
+            },
+            pageCount: 304,
+            printType: "BOOK",
+            categories: ["Fiction"],
+            averageRating: 3.5,
+            ratingsCount: 10,
+            maturityRating: "NOT_MATURE",
+            allowAnonLogging: false,
+            contentVersion: "1.2.3.0.preview.2",
+            panelizationSummary: {
+              containsEpubBubbles: false,
+              containsImageBubbles: false,
+            },
+            imageLinks: {
+              smallThumbnail:
+                "http://books.google.com/books/content?id=hqsJKvM17WwC&printsec=frontcover&img=1&zoom=5&edge=curl&source=gbs_api",
+              thumbnail:
+                "http://books.google.com/books/content?id=hqsJKvM17WwC&printsec=frontcover&img=1&zoom=1&edge=curl&source=gbs_api",
+            },
+            language: "en",
+            previewLink:
+              "http://books.google.co.uk/books?id=hqsJKvM17WwC&printsec=frontcover&dq=isbn:9780143119111&hl=&cd=1&source=gbs_api",
+            infoLink:
+              "https://play.google.com/store/books/details?id=hqsJKvM17WwC&source=gbs_api",
+            canonicalVolumeLink:
+              "https://play.google.com/store/books/details?id=hqsJKvM17WwC",
+          },
+          saleInfo: {
+            country: "GB",
+            saleability: "FOR_SALE",
+            isEbook: true,
+            listPrice: {
+              amount: 4.23,
+              currencyCode: "GBP",
+            },
+            retailPrice: {
+              amount: 4.23,
+              currencyCode: "GBP",
+            },
+            buyLink:
+              "https://play.google.com/store/books/details?id=hqsJKvM17WwC&rdid=book-hqsJKvM17WwC&rdot=1&source=gbs_api",
+            offers: [
+              {
+                finskyOfferType: 1,
+                listPrice: {
+                  amountInMicros: 4230000,
+                  currencyCode: "GBP",
+                },
+                retailPrice: {
+                  amountInMicros: 4230000,
+                  currencyCode: "GBP",
+                },
+                giftable: true,
+              },
+            ],
+          },
+          accessInfo: {
+            country: "GB",
+            viewability: "PARTIAL",
+            embeddable: true,
+            publicDomain: false,
+            textToSpeechPermission: "ALLOWED",
+            epub: {
+              isAvailable: true,
+              acsTokenLink:
+                "http://books.google.co.uk/books/download/Gulliver_s_Travels-sample-epub.acsm?id=hqsJKvM17WwC&format=epub&output=acs4_fulfillment_token&dl_type=sample&source=gbs_api",
+            },
+            pdf: {
+              isAvailable: false,
+            },
+            webReaderLink:
+              "http://play.google.com/books/reader?id=hqsJKvM17WwC&hl=&printsec=frontcover&source=gbs_api",
+            accessViewStatus: "SAMPLE",
+            quoteSharingAllowed: false,
+          },
+          searchInfo: {
+            textSnippet:
+              "The series was founded in 1917 by the publishers Boni and Liveright and eight years later acquired by Bennett Cerf and Donald Klopfer. It provided the foundation for their next publishing venture, Random House.",
+          },
+        },
+      ],
+    },
+    {
+      kind: "books#volumes",
+      totalItems: 1,
+      items: [
+        {
+          kind: "books#volume",
+          id: "N6DeDQAAQBAJ",
+          etag: "U0SEA/YgPNk",
+          selfLink: "https://www.googleapis.com/books/v1/volumes/N6DeDQAAQBAJ",
+          volumeInfo: {
+            title: "Harry Potter and the Deathly Hallows",
+            authors: ["J. K. Rowling"],
+            publisher: "Bloomsbury Publishing",
+            publishedDate: "2015-08-13",
+            description:
+              "Harry Potter is preparing to leave the Dursleys and Privet Drive for the last time. But the future that awaits him is full of danger, not only for him, but for anyone close to him - and Harry has already lost so much. Only by destroying Voldemort's remaining Horcruxes can Harry free himself and overcome the Dark Lord's forces of evil. In this dramatic conclusion to the Harry Potter series, Harry must leave his most loyal friends behind, and in a final perilous journey find the strength and the will to face his terrifying destiny: a deadly confrontation that is his alone to fight. These adult editions with glorious jacket art by Andrew Davidson are now available in hardback for the first time.",
+            industryIdentifiers: [
+              {
+                type: "ISBN_13",
+                identifier: "9781408865453",
+              },
+              {
+                type: "ISBN_10",
+                identifier: "1408865459",
+              },
+            ],
+            readingModes: {
+              text: false,
+              image: true,
+            },
+            pageCount: 619,
+            printType: "BOOK",
+            categories: ["Boarding schools"],
+            averageRating: 4.5,
+            ratingsCount: 3382,
+            maturityRating: "NOT_MATURE",
+            allowAnonLogging: false,
+            contentVersion: "0.1.0.0.preview.1",
+            panelizationSummary: {
+              containsEpubBubbles: false,
+              containsImageBubbles: false,
+            },
+            imageLinks: {
+              smallThumbnail:
+                "http://books.google.com/books/content?id=N6DeDQAAQBAJ&printsec=frontcover&img=1&zoom=5&edge=curl&source=gbs_api",
+              thumbnail:
+                "http://books.google.com/books/content?id=N6DeDQAAQBAJ&printsec=frontcover&img=1&zoom=1&edge=curl&source=gbs_api",
+            },
+            language: "en",
+            previewLink:
+              "http://books.google.co.uk/books?id=N6DeDQAAQBAJ&printsec=frontcover&dq=isbn:9781408865453&hl=&cd=1&source=gbs_api",
+            infoLink:
+              "http://books.google.co.uk/books?id=N6DeDQAAQBAJ&dq=isbn:9781408865453&hl=&source=gbs_api",
+            canonicalVolumeLink:
+              "https://books.google.com/books/about/Harry_Potter_and_the_Deathly_Hallows.html?hl=&id=N6DeDQAAQBAJ",
+          },
+          saleInfo: {
+            country: "GB",
+            saleability: "NOT_FOR_SALE",
+            isEbook: false,
+          },
+          accessInfo: {
+            country: "GB",
+            viewability: "PARTIAL",
+            embeddable: true,
+            publicDomain: false,
+            textToSpeechPermission: "ALLOWED",
+            epub: {
+              isAvailable: false,
+            },
+            pdf: {
+              isAvailable: false,
+            },
+            webReaderLink:
+              "http://play.google.com/books/reader?id=N6DeDQAAQBAJ&hl=&printsec=frontcover&source=gbs_api",
+            accessViewStatus: "SAMPLE",
+            quoteSharingAllowed: false,
+          },
+          searchInfo: {
+            textSnippet:
+              "Harry has been burdened with a dark, dangerous and seemingly impossible task - that of locating and destroying Voldemort&#39;s remaining Horcruxes.",
+          },
+        },
+      ],
+    },
+  ];
+
+export default books;
